Working on implementing folding graphics as a function into the main program - still need to sort crashing bug with instruction reg overwrite but may be an artifact of implementation with folding_graphics.c
